+++ lib/go/text_client/state.ex
@@ -1,0 +1,9 @@
+defmodule Go.TextClient.State do
+  @moduledoc false
+    
+  defstruct(
+    game_service: nil,
+    tally: nil,
+    move: nil
+  )
+end