+++ lib/go/text_client/text_client.ex
@@ -1,0 +1,7 @@
+defmodule Go.TextClient do
+  @moduledoc false
+  
+  alias Go.TextClient.Interact
+  
+  defdelegate start(size \\ 19), to: Interact
+end